+++ app/posts/[Id]/postdetail/page.tsx
@@ -10,11 +10,11 @@ import CommentForm from "@/app/components/comments/CommentForm";
 dayjs.extend(relativeTime);
 
 type PostDetailPageProps = {
-  params: { id: string };
+  params: Promise<{ id: string }>
 };
 
 export default async function PostDetailPage({ params }: PostDetailPageProps) {
-  const { id } = await params;
+  const { id } = (await params) ?? {};
   const postId = Number(id);
 
   const postDetail = await getPostDetail(id);
@@ -83,3 +83,4 @@ export default async function PostDetailPage({ params }: PostDetailPageProps) {
     </>
   );
 }
+
