+++ app/posts/create/page.tsx
@@ -13,16 +13,7 @@ import { useForm } from "react-hook-form";
 import { z } from "zod";
 import { useRouter } from "next/navigation";
 import { isAuthenticated } from "@/lib/api/auth";
-
-export const formSchema = z.object({
-  title: z.string().min(2, { message: "タイトルは2文字以上で入力してください。" }),
-  images: z.object({ url: z.string() }).array(),
-  category: z.enum(["value"], { message: "カテゴリーを選択してください。" }),
-  content: z
-    .string()
-    .min(10, { message: "本文は10文字以上で入力してください。" })
-    .max(1000, { message: "本文は1000文字以内で入力してください。" }),
-});
+import { formSchema } from "@/lib/posts/createFormSchema";
 
 const CreateBBSPage = () => {
   const router = useRouter();
